pyproject.toml
[tool.poetry]
packages = [
  { include = "competencyAnalyser" },
  { include = "telegram_bot" },
]
name = "competencyAnalyser"
version = "0.3.0"
description = ""
authors = ["skip"]
readme="README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
jinja2 = "^3.1.4"
python-multipart = "^0.0.9"
authlib = "^1.3.1"
sqlalchemy = "^2.0.31"
itsdangerous = "^2.2.0"
alembic = "^1.13.1"
requests = "^2.32.3"
openai = "^1.35.3"
httpx = "^0.23.0"
regex = "^2024.5.15"
google = "^3.0.0"
google-auth = "^2.31.0"
hcaptcha = "^0.2.1"
insecure-package = "^0.1.0"
telebot = "^0.0.5"
psycopg2-binary = "^2.9.9"
telegram-webapp-auth = "^1.0.1"
translate = "^3.6.1"
googletrans-py = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"

[tool.poetry.scripts]
competency-analyser = "competencyAnalyser.scripts.competencyAnalyser:start"
competency-analyser-telegram-bot = "telegram_bot.telegram_bot:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
